Inline compat.py
