+++ scripts/create_manifest.py
@@ -1,10 +1,24 @@
 import argparse
-import dxpy
-from dxpy.utils.resolver import resolve_existing_path
+import bz2
+import json
+import sys
 import collections
 import os
+import dxpy
+from dxpy.utils.resolver import resolve_existing_path
 
-import compat
+# Provide compatibility between python 2 and 3
+USING_PYTHON2 = True if sys.version_info < (3, 0) else False
+
+def write_manifest_to_file(outfile, manifest):
+    if USING_PYTHON2:
+        with open(outfile, "w") as f:
+            f.write(bz2.compress(json.dumps(manifest, indent=2, sort_keys=True)))
+    else:
+        # python 3 requires opening the file in binary mode
+        with open(outfile, "wb") as f:
+            value = json.dumps(manifest, indent=2, sort_keys=True)
+            f.write(bz2.compress(value.encode()))
 
 def fileID2manifest(fdetails, project):
     """
@@ -49,7 +63,7 @@ def main():
             fname, fext = os.path.splitext(x['name'])
             x['name'] = fname + "_" + x['id'] + fext
 
-    compat.write_manifest_to_file(args.output_file, manifest)
+    write_manifest_to_file(args.output_file, manifest)
     print("Manifest file written to {}".format(args.output_file))
     print("Total {} objects".format(len(manifest[project])))
